+++ packages/amqp/index.js
@@ -1,5 +1,6 @@
 import { connect } from './lib/connect.js';
 import { consumeFactory } from './lib/consume.js';
+import { createConsumer } from './lib/consumer.js';
 import { createPublisher } from './lib/publisher.js';
 import { publishFactory } from './lib/publish.js';
 
@@ -20,37 +21,4 @@ export const createClient = async ({
     };
 };
 
-export const createConsumer = async ({
-    mqUrl,
-    tlsOptions,
-    consumer,
-    durable = true,
-    exchange,
-    name,
-    noAck = false,
-    queue,
-    routingKey,
-    type = 'topic',
-    ...sessionOptions
-}) => {
-    const session = await connect({ mqUrl, tlsOptions, ...sessionOptions });
-    const { consume } = consumeFactory({ session });
-
-    const result = await consume({
-        consumer,
-        durable,
-        exchange,
-        name,
-        noAck,
-        queue,
-        routingKey,
-        type,
-    });
-
-    return {
-        ...result,
-        stop: (reason) => session.stop(reason),
-    };
-};
-
-export { createPublisher };
+export { createConsumer, createPublisher };
