+++ packages/amqp/lib/publisher.js
@@ -5,11 +5,11 @@ import { connect } from './connect.js';
 const multiLog = createMultiLog();
 
 export const createPublisher = async ({
-    mqUrl,
-    tlsOptions,
     durable = true,
     exchange,
+    mqUrl,
     routingKey,
+    tlsOptions,
     type = 'topic',
     ...sessionOptions
 }) => {
@@ -24,19 +24,22 @@ export const createPublisher = async ({
     const session = await connect({ mqUrl, tlsOptions, ...sessionOptions });
     const x = await session.exchange(exchange, type, { durable });
 
-    return async ({ data }) => {
-        if (data === undefined) {
-            throw new Error('The data parameter must be provided');
-        }
+    return {
+        publish: async ({ data }) => {
+            if (data === undefined) {
+                throw new Error('The data parameter must be provided');
+            }
 
-        multiLog(
-            {
-                debug: { data, type },
-                info: { exchange, routingKey },
-            },
-            'Publishing a message',
-        );
+            multiLog(
+                {
+                    debug: { data, type },
+                    info: { exchange, routingKey },
+                },
+                'Publishing a message',
+            );
 
-        await x.publish(data, { routingKey });
+            await x.publish(data, { routingKey });
+        },
+        stop: (reason) => session.stop(reason),
     };
 };
